Move list functionality into gorm package
