+++ server/registry/internal/storage/artifacts.go
@@ -17,318 +17,29 @@ package storage
 import (
 	"context"
 
-	"github.com/apigee/registry/server/registry/internal/storage/filtering"
 	"github.com/apigee/registry/server/registry/internal/storage/gorm"
 	"github.com/apigee/registry/server/registry/internal/storage/models"
 	"github.com/apigee/registry/server/registry/names"
-	"google.golang.org/api/iterator"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
-// ArtifactList contains a page of artifact resources.
-type ArtifactList struct {
-	Artifacts []models.Artifact
-	Token     string
+func (d *Client) ListSpecArtifacts(ctx context.Context, parent names.Spec, opts gorm.PageOptions) (gorm.ArtifactList, error) {
+	return d.Client.ListSpecArtifacts(ctx, parent, opts)
 }
 
-var artifactFields = []filtering.Field{
-	{Name: "name", Type: filtering.String},
-	{Name: "project_id", Type: filtering.String},
-	{Name: "api_id", Type: filtering.String},
-	{Name: "version_id", Type: filtering.String},
-	{Name: "spec_id", Type: filtering.String},
-	{Name: "artifact_id", Type: filtering.String},
-	{Name: "create_time", Type: filtering.Timestamp},
-	{Name: "update_time", Type: filtering.Timestamp},
-	{Name: "mime_type", Type: filtering.String},
-	{Name: "size_bytes", Type: filtering.Int},
+func (d *Client) ListVersionArtifacts(ctx context.Context, parent names.Version, opts gorm.PageOptions) (gorm.ArtifactList, error) {
+	return d.Client.ListVersionArtifacts(ctx, parent, opts)
 }
 
-func (d *Client) ListSpecArtifacts(ctx context.Context, parent names.Spec, opts PageOptions) (ArtifactList, error) {
-	q := d.NewQuery(gorm.ArtifactEntityName)
-	q = q.Require("DeploymentID", "")
-
-	token, err := decodeToken(opts.Token)
-	if err != nil {
-		return ArtifactList{}, status.Errorf(codes.InvalidArgument, "invalid page token %q: %s", opts.Token, err.Error())
-	}
-
-	if err := token.ValidateFilter(opts.Filter); err != nil {
-		return ArtifactList{}, status.Errorf(codes.InvalidArgument, "invalid filter %q: %s", opts.Filter, err)
-	} else {
-		token.Filter = opts.Filter
-	}
-
-	q = q.ApplyOffset(token.Offset)
-
-	if id := parent.ProjectID; id != "-" {
-		q = q.Require("ProjectID", id)
-	}
-	if id := parent.ApiID; id != "-" {
-		q = q.Require("ApiID", id)
-	}
-	if id := parent.VersionID; id != "-" {
-		q = q.Require("VersionID", id)
-	}
-	if id := parent.SpecID; id != "-" {
-		q = q.Require("SpecID", id)
-	}
-
-	if parent.ProjectID != "-" && parent.ApiID != "-" && parent.VersionID != "-" && parent.SpecID != "-" {
-		if _, err := d.GetSpec(ctx, parent); err != nil {
-			return ArtifactList{}, err
-		}
-	} else if parent.ProjectID != "-" && parent.ApiID != "-" && parent.VersionID != "-" && parent.SpecID == "-" {
-		if _, err := d.GetVersion(ctx, parent.Version()); err != nil {
-			return ArtifactList{}, err
-		}
-	} else if parent.ProjectID != "-" && parent.ApiID != "-" && parent.VersionID == "-" && parent.SpecID == "-" {
-		if _, err := d.GetApi(ctx, parent.Api()); err != nil {
-			return ArtifactList{}, err
-		}
-	} else if parent.ProjectID != "-" && parent.ApiID == "-" && parent.VersionID == "-" && parent.SpecID == "-" {
-		if _, err := d.GetProject(ctx, parent.Project()); err != nil {
-			return ArtifactList{}, err
-		}
-	}
-
-	return d.listArtifacts(ctx, d.Run(ctx, q), opts, func(a *models.Artifact) bool {
-		return a.ProjectID != "" && a.ApiID != "" && a.VersionID != "" && a.SpecID != ""
-	})
+func (d *Client) ListDeploymentArtifacts(ctx context.Context, parent names.Deployment, opts gorm.PageOptions) (gorm.ArtifactList, error) {
+	return d.Client.ListDeploymentArtifacts(ctx, parent, opts)
 }
 
-func (d *Client) ListVersionArtifacts(ctx context.Context, parent names.Version, opts PageOptions) (ArtifactList, error) {
-	q := d.NewQuery(gorm.ArtifactEntityName)
-	q = q.Require("SpecID", "")
-	q = q.Require("DeploymentID", "")
-
-	token, err := decodeToken(opts.Token)
-	if err != nil {
-		return ArtifactList{}, status.Errorf(codes.InvalidArgument, "invalid page token %q: %s", opts.Token, err.Error())
-	}
-
-	if err := token.ValidateFilter(opts.Filter); err != nil {
-		return ArtifactList{}, status.Errorf(codes.InvalidArgument, "invalid filter %q: %s", opts.Filter, err)
-	} else {
-		token.Filter = opts.Filter
-	}
-
-	q = q.ApplyOffset(token.Offset)
-
-	if id := parent.ProjectID; id != "-" {
-		q = q.Require("ProjectID", id)
-	}
-	if id := parent.ApiID; id != "-" {
-		q = q.Require("ApiID", id)
-	}
-	if id := parent.VersionID; id != "-" {
-		q = q.Require("VersionID", id)
-	}
-
-	if parent.ProjectID != "-" && parent.ApiID != "-" && parent.VersionID != "-" {
-		if _, err := d.GetVersion(ctx, parent); err != nil {
-			return ArtifactList{}, err
-		}
-	} else if parent.ProjectID != "-" && parent.ApiID != "-" && parent.VersionID == "-" {
-		if _, err := d.GetApi(ctx, parent.Api()); err != nil {
-			return ArtifactList{}, err
-		}
-	} else if parent.ProjectID != "-" && parent.ApiID == "-" && parent.VersionID == "-" {
-		if _, err := d.GetProject(ctx, parent.Project()); err != nil {
-			return ArtifactList{}, err
-		}
-	}
-
-	return d.listArtifacts(ctx, d.Run(ctx, q), opts, func(a *models.Artifact) bool {
-		return a.ProjectID != "" && a.ApiID != "" && a.VersionID != ""
-	})
+func (d *Client) ListApiArtifacts(ctx context.Context, parent names.Api, opts gorm.PageOptions) (gorm.ArtifactList, error) {
+	return d.Client.ListApiArtifacts(ctx, parent, opts)
 }
 
-func (d *Client) ListDeploymentArtifacts(ctx context.Context, parent names.Deployment, opts PageOptions) (ArtifactList, error) {
-	q := d.NewQuery(gorm.ArtifactEntityName)
-	q = q.Require("SpecID", "")
-	q = q.Require("VersionID", "")
-
-	token, err := decodeToken(opts.Token)
-	if err != nil {
-		return ArtifactList{}, status.Errorf(codes.InvalidArgument, "invalid page token %q: %s", opts.Token, err.Error())
-	}
-
-	if err := token.ValidateFilter(opts.Filter); err != nil {
-		return ArtifactList{}, status.Errorf(codes.InvalidArgument, "invalid filter %q: %s", opts.Filter, err)
-	} else {
-		token.Filter = opts.Filter
-	}
-
-	q = q.ApplyOffset(token.Offset)
-
-	if id := parent.ProjectID; id != "-" {
-		q = q.Require("ProjectID", id)
-	}
-	if id := parent.ApiID; id != "-" {
-		q = q.Require("ApiID", id)
-	}
-	if id := parent.DeploymentID; id != "-" {
-		q = q.Require("DeploymentID", id)
-	}
-
-	if parent.ProjectID != "-" && parent.ApiID != "-" && parent.DeploymentID != "-" {
-		if _, err := d.GetDeployment(ctx, parent); err != nil {
-			return ArtifactList{}, err
-		}
-	} else if parent.ProjectID != "-" && parent.ApiID != "-" && parent.DeploymentID == "-" {
-		if _, err := d.GetApi(ctx, parent.Api()); err != nil {
-			return ArtifactList{}, err
-		}
-	} else if parent.ProjectID != "-" && parent.ApiID == "-" && parent.DeploymentID == "-" {
-		if _, err := d.GetProject(ctx, parent.Project()); err != nil {
-			return ArtifactList{}, err
-		}
-	}
-
-	return d.listArtifacts(ctx, d.Run(ctx, q), opts, func(a *models.Artifact) bool {
-		return a.ProjectID != "" && a.ApiID != "" && a.DeploymentID != ""
-	})
-}
-
-func (d *Client) ListApiArtifacts(ctx context.Context, parent names.Api, opts PageOptions) (ArtifactList, error) {
-	q := d.NewQuery(gorm.ArtifactEntityName)
-	q = q.Require("DeploymentID", "")
-	q = q.Require("VersionID", "")
-	q = q.Require("SpecID", "")
-
-	token, err := decodeToken(opts.Token)
-	if err != nil {
-		return ArtifactList{}, status.Errorf(codes.InvalidArgument, "invalid page token %q: %s", opts.Token, err.Error())
-	}
-
-	if err := token.ValidateFilter(opts.Filter); err != nil {
-		return ArtifactList{}, status.Errorf(codes.InvalidArgument, "invalid filter %q: %s", opts.Filter, err)
-	} else {
-		token.Filter = opts.Filter
-	}
-
-	q = q.ApplyOffset(token.Offset)
-
-	if id := parent.ProjectID; id != "-" {
-		q = q.Require("ProjectID", id)
-	}
-	if id := parent.ApiID; id != "-" {
-		q = q.Require("ApiID", id)
-	}
-
-	if parent.ProjectID != "-" && parent.ApiID != "-" {
-		if _, err := d.GetApi(ctx, parent); err != nil {
-			return ArtifactList{}, err
-		}
-	} else if parent.ProjectID != "-" && parent.ApiID == "-" {
-		if _, err := d.GetProject(ctx, parent.Project()); err != nil {
-			return ArtifactList{}, err
-		}
-	}
-
-	return d.listArtifacts(ctx, d.Run(ctx, q), opts, func(a *models.Artifact) bool {
-		return a.ProjectID != "" && a.ApiID != ""
-	})
-}
-
-func (d *Client) ListProjectArtifacts(ctx context.Context, parent names.Project, opts PageOptions) (ArtifactList, error) {
-	q := d.NewQuery(gorm.ArtifactEntityName)
-	q = q.Require("ApiID", "")
-	q = q.Require("DeploymentID", "")
-	q = q.Require("VersionID", "")
-	q = q.Require("SpecID", "")
-
-	token, err := decodeToken(opts.Token)
-	if err != nil {
-		return ArtifactList{}, status.Errorf(codes.InvalidArgument, "invalid page token %q: %s", opts.Token, err.Error())
-	}
-
-	if err := token.ValidateFilter(opts.Filter); err != nil {
-		return ArtifactList{}, status.Errorf(codes.InvalidArgument, "invalid filter %q: %s", opts.Filter, err)
-	} else {
-		token.Filter = opts.Filter
-	}
-
-	q = q.ApplyOffset(token.Offset)
-
-	if id := parent.ProjectID; id != "-" {
-		q = q.Require("ProjectID", id)
-		if _, err := d.GetProject(ctx, parent); err != nil {
-			return ArtifactList{}, err
-		}
-	}
-
-	return d.listArtifacts(ctx, d.Run(ctx, q), opts, func(a *models.Artifact) bool {
-		return a.ProjectID != ""
-	})
-}
-
-func (d *Client) listArtifacts(ctx context.Context, it *gorm.Iterator, opts PageOptions, include func(*models.Artifact) bool) (ArtifactList, error) {
-	token, err := decodeToken(opts.Token)
-	if err != nil {
-		return ArtifactList{}, status.Errorf(codes.InvalidArgument, "invalid page token %q: %s", opts.Token, err.Error())
-	} else {
-		token.Filter = opts.Filter
-	}
-
-	filter, err := filtering.NewFilter(opts.Filter, artifactFields)
-	if err != nil {
-		return ArtifactList{}, err
-	}
-
-	response := ArtifactList{
-		Artifacts: make([]models.Artifact, 0, opts.Size),
-	}
-
-	artifact := new(models.Artifact)
-	for _, err = it.Next(artifact); err == nil; _, err = it.Next(artifact) {
-		artifactMap, err := artifactMap(*artifact)
-		if err != nil {
-			return response, status.Error(codes.Internal, err.Error())
-		}
-
-		match, err := filter.Matches(artifactMap)
-		if err != nil {
-			return response, err
-		} else if !match || !include(artifact) {
-			token.Offset++
-			continue
-		} else if len(response.Artifacts) == int(opts.Size) {
-			break
-		}
-
-		response.Artifacts = append(response.Artifacts, *artifact)
-		token.Offset++
-	}
-	if err != nil && err != iterator.Done {
-		return response, status.Error(codes.Internal, err.Error())
-	}
-
-	if err == nil {
-		response.Token, err = encodeToken(token)
-		if err != nil {
-			return response, status.Error(codes.Internal, err.Error())
-		}
-	}
-
-	return response, nil
-}
-
-func artifactMap(artifact models.Artifact) (map[string]interface{}, error) {
-	return map[string]interface{}{
-		"name":        artifact.Name(),
-		"project_id":  artifact.ProjectID,
-		"api_id":      artifact.ApiID,
-		"version_id":  artifact.VersionID,
-		"spec_id":     artifact.SpecID,
-		"artifact_id": artifact.ArtifactID,
-		"create_time": artifact.CreateTime,
-		"update_time": artifact.UpdateTime,
-		"mime_type":   artifact.MimeType,
-		"size_bytes":  artifact.SizeInBytes,
-	}, nil
+func (d *Client) ListProjectArtifacts(ctx context.Context, parent names.Project, opts gorm.PageOptions) (gorm.ArtifactList, error) {
+	return d.Client.ListProjectArtifacts(ctx, parent, opts)
 }
 
 func (d *Client) SaveArtifact(ctx context.Context, artifact *models.Artifact) error {
