+++ server/registry/internal/storage/deployment_revisions.go
@@ -20,55 +20,10 @@ import (
 	"github.com/apigee/registry/server/registry/internal/storage/gorm"
 	"github.com/apigee/registry/server/registry/internal/storage/models"
 	"github.com/apigee/registry/server/registry/names"
-	"google.golang.org/api/iterator"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
-func (d *Client) ListDeploymentRevisions(ctx context.Context, parent names.Deployment, opts PageOptions) (DeploymentList, error) {
-	q := d.NewQuery(gorm.DeploymentEntityName)
-	q = q.Require("ProjectID", parent.ProjectID)
-	q = q.Require("ApiID", parent.ApiID)
-	q = q.Require("DeploymentID", parent.DeploymentID)
-	q = q.Descending("RevisionCreateTime")
-
-	token, err := decodeToken(opts.Token)
-	if err != nil {
-		return DeploymentList{}, status.Errorf(codes.InvalidArgument, "invalid page token %q: %s", opts.Token, err.Error())
-	}
-
-	if err := token.ValidateFilter(opts.Filter); err != nil {
-		return DeploymentList{}, status.Errorf(codes.InvalidArgument, "invalid filter %q: %s", opts.Filter, err)
-	}
-
-	q = q.ApplyOffset(token.Offset)
-
-	it := d.Run(ctx, q)
-	response := DeploymentList{
-		Deployments: make([]models.Deployment, 0, opts.Size),
-	}
-
-	revision := new(models.Deployment)
-	for _, err = it.Next(revision); err == nil; _, err = it.Next(revision) {
-		token.Offset++
-
-		response.Deployments = append(response.Deployments, *revision)
-		if len(response.Deployments) == int(opts.Size) {
-			break
-		}
-	}
-	if err != nil && err != iterator.Done {
-		return response, status.Error(codes.Internal, err.Error())
-	}
-
-	if err == nil {
-		response.Token, err = encodeToken(token)
-		if err != nil {
-			return response, status.Error(codes.Internal, err.Error())
-		}
-	}
-
-	return response, nil
+func (d *Client) ListDeploymentRevisions(ctx context.Context, parent names.Deployment, opts gorm.PageOptions) (gorm.DeploymentList, error) {
+	return d.Client.ListDeploymentRevisions(ctx, parent, opts)
 }
 
 func (d *Client) SaveDeploymentRevision(ctx context.Context, revision *models.Deployment) error {
