+++ server/registry/internal/storage/gorm/client.go
@@ -263,73 +263,6 @@ func (c *Client) Run(ctx context.Context, q *Query) *Iterator {
 	}
 }
 
-func (c *Client) GetRecentSpecRevisions(ctx context.Context, offset int32, projectID, apiID, versionID string) *Iterator {
-	lock()
-	defer unlock()
-
-	// Select all columns from `specs` table specifically.
-	// We do not want to select duplicates from the joined subquery result.
-	op := c.db.Select("specs.*").
-		Table("specs").
-		// Join missing columns that couldn't be selected in the subquery.
-		Joins("JOIN (?) AS grp ON specs.project_id = grp.project_id AND specs.api_id = grp.api_id AND specs.version_id = grp.version_id AND specs.spec_id = grp.spec_id AND specs.revision_create_time = grp.recent_create_time",
-			// Select spec names and only their most recent revision_create_time
-			// This query cannot select all the columns we want.
-			// See: https://stackoverflow.com/questions/7745609/sql-select-only-rows-with-max-value-on-a-column
-			c.db.Select("project_id, api_id, version_id, spec_id, MAX(revision_create_time) AS recent_create_time").
-				Table("specs").
-				Group("project_id, api_id, version_id, spec_id")).
-		Order("key").
-		Offset(int(offset)).
-		Limit(100000)
-
-	if projectID != "-" {
-		op = op.Where("specs.project_id = ?", projectID)
-	}
-	if apiID != "-" {
-		op = op.Where("specs.api_id = ?", apiID)
-	}
-	if versionID != "-" {
-		op = op.Where("specs.version_id = ?", versionID)
-	}
-
-	var v []models.Spec
-	_ = op.Scan(&v).Error
-	return &Iterator{Values: v, Index: 0}
-}
-
-func (c *Client) GetRecentDeploymentRevisions(ctx context.Context, offset int32, projectID, apiID string) *Iterator {
-	lock()
-	defer unlock()
-
-	// Select all columns from `deployments` table specifically.
-	// We do not want to select duplicates from the joined subquery result.
-	op := c.db.Select("deployments.*").
-		Table("deployments").
-		// Join missing columns that couldn't be selected in the subquery.
-		Joins("JOIN (?) AS grp ON deployments.project_id = grp.project_id AND deployments.api_id = grp.api_id AND deployments.deployment_id = grp.deployment_id AND deployments.revision_create_time = grp.recent_create_time",
-			// Select deployment names and only their most recent revision_create_time
-			// This query cannot select all the columns we want.
-			// See: https://stackoverflow.com/questions/7745609/sql-select-only-rows-with-max-value-on-a-column
-			c.db.Select("project_id, api_id, deployment_id, MAX(revision_create_time) AS recent_create_time").
-				Table("deployments").
-				Group("project_id, api_id, deployment_id")).
-		Order("key").
-		Offset(int(offset)).
-		Limit(100000)
-
-	if projectID != "-" {
-		op = op.Where("deployments.project_id = ?", projectID)
-	}
-	if apiID != "-" {
-		op = op.Where("deployments.api_id = ?", apiID)
-	}
-
-	var v []models.Deployment
-	_ = op.Scan(&v).Error
-	return &Iterator{Values: v, Index: 0}
-}
-
 func (c *Client) Migrate(kind string) error {
 	return c.db.AutoMigrate(entities...)
 }
