+++ server/registry/internal/storage/spec_revisions.go
@@ -20,56 +20,10 @@ import (
 	"github.com/apigee/registry/server/registry/internal/storage/gorm"
 	"github.com/apigee/registry/server/registry/internal/storage/models"
 	"github.com/apigee/registry/server/registry/names"
-	"google.golang.org/api/iterator"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
-func (d *Client) ListSpecRevisions(ctx context.Context, parent names.Spec, opts PageOptions) (SpecList, error) {
-	q := d.NewQuery(gorm.SpecEntityName)
-	q = q.Require("ProjectID", parent.ProjectID)
-	q = q.Require("ApiID", parent.ApiID)
-	q = q.Require("VersionID", parent.VersionID)
-	q = q.Require("SpecID", parent.SpecID)
-	q = q.Descending("RevisionCreateTime")
-
-	token, err := decodeToken(opts.Token)
-	if err != nil {
-		return SpecList{}, status.Errorf(codes.InvalidArgument, "invalid page token %q: %s", opts.Token, err.Error())
-	}
-
-	if err := token.ValidateFilter(opts.Filter); err != nil {
-		return SpecList{}, status.Errorf(codes.InvalidArgument, "invalid filter %q: %s", opts.Filter, err)
-	}
-
-	q = q.ApplyOffset(token.Offset)
-
-	it := d.Run(ctx, q)
-	response := SpecList{
-		Specs: make([]models.Spec, 0, opts.Size),
-	}
-
-	revision := new(models.Spec)
-	for _, err = it.Next(revision); err == nil; _, err = it.Next(revision) {
-		token.Offset++
-
-		response.Specs = append(response.Specs, *revision)
-		if len(response.Specs) == int(opts.Size) {
-			break
-		}
-	}
-	if err != nil && err != iterator.Done {
-		return response, status.Error(codes.Internal, err.Error())
-	}
-
-	if err == nil {
-		response.Token, err = encodeToken(token)
-		if err != nil {
-			return response, status.Error(codes.Internal, err.Error())
-		}
-	}
-
-	return response, nil
+func (d *Client) ListSpecRevisions(ctx context.Context, parent names.Spec, opts gorm.PageOptions) (gorm.SpecList, error) {
+	return d.Client.ListSpecRevisions(ctx, parent, opts)
 }
 
 func (d *Client) SaveSpecRevision(ctx context.Context, revision *models.Spec) error {
