+++ server/registry/internal/storage/specs.go
@@ -17,132 +17,13 @@ package storage
 import (
 	"context"
 
-	"github.com/apigee/registry/server/registry/internal/storage/filtering"
 	"github.com/apigee/registry/server/registry/internal/storage/gorm"
 	"github.com/apigee/registry/server/registry/internal/storage/models"
 	"github.com/apigee/registry/server/registry/names"
-	"google.golang.org/api/iterator"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
-// SpecList contains a page of spec resources.
-type SpecList struct {
-	Specs []models.Spec
-	Token string
-}
-
-var specFields = []filtering.Field{
-	{Name: "name", Type: filtering.String},
-	{Name: "project_id", Type: filtering.String},
-	{Name: "api_id", Type: filtering.String},
-	{Name: "version_id", Type: filtering.String},
-	{Name: "spec_id", Type: filtering.String},
-	{Name: "filename", Type: filtering.String},
-	{Name: "description", Type: filtering.String},
-	{Name: "create_time", Type: filtering.Timestamp},
-	{Name: "revision_create_time", Type: filtering.Timestamp},
-	{Name: "revision_update_time", Type: filtering.Timestamp},
-	{Name: "mime_type", Type: filtering.String},
-	{Name: "size_bytes", Type: filtering.Int},
-	{Name: "source_uri", Type: filtering.String},
-	{Name: "labels", Type: filtering.StringMap},
-}
-
-func (d *Client) ListSpecs(ctx context.Context, parent names.Version, opts PageOptions) (SpecList, error) {
-	token, err := decodeToken(opts.Token)
-	if err != nil {
-		return SpecList{}, status.Errorf(codes.InvalidArgument, "invalid page token %q: %s", opts.Token, err.Error())
-	}
-
-	if err := token.ValidateFilter(opts.Filter); err != nil {
-		return SpecList{}, status.Errorf(codes.InvalidArgument, "invalid filter %q: %s", opts.Filter, err)
-	} else {
-		token.Filter = opts.Filter
-	}
-
-	if parent.ProjectID != "-" && parent.ApiID != "-" && parent.VersionID != "-" {
-		if _, err := d.GetVersion(ctx, parent); err != nil {
-			return SpecList{}, err
-		}
-	} else if parent.ProjectID != "-" && parent.ApiID != "-" && parent.VersionID == "-" {
-		if _, err := d.GetApi(ctx, parent.Api()); err != nil {
-			return SpecList{}, err
-		}
-	} else if parent.ProjectID != "-" && parent.ApiID == "-" && parent.VersionID == "-" {
-		if _, err := d.GetProject(ctx, parent.Project()); err != nil {
-			return SpecList{}, err
-		}
-	}
-
-	filter, err := filtering.NewFilter(opts.Filter, specFields)
-	if err != nil {
-		return SpecList{}, err
-	}
-
-	it := d.GetRecentSpecRevisions(ctx, token.Offset, parent.ProjectID, parent.ApiID, parent.VersionID)
-	response := SpecList{
-		Specs: make([]models.Spec, 0, opts.Size),
-	}
-
-	spec := new(models.Spec)
-	for _, err = it.Next(spec); err == nil; _, err = it.Next(spec) {
-		specMap, err := specMap(*spec)
-		if err != nil {
-			return response, status.Error(codes.Internal, err.Error())
-		}
-
-		match, err := filter.Matches(specMap)
-		if err != nil {
-			return response, err
-		} else if !match {
-			token.Offset++
-			continue
-		} else if len(response.Specs) == int(opts.Size) {
-			break
-		}
-
-		response.Specs = append(response.Specs, *spec)
-		token.Offset++
-	}
-	if err != nil && err != iterator.Done {
-		return response, status.Error(codes.Internal, err.Error())
-	}
-
-	if err == nil {
-		response.Token, err = encodeToken(token)
-		if err != nil {
-			return response, status.Error(codes.Internal, err.Error())
-		}
-	}
-
-	return response, nil
-}
-
-func specMap(spec models.Spec) (map[string]interface{}, error) {
-	labels, err := spec.LabelsMap()
-	if err != nil {
-		return nil, err
-	}
-
-	return map[string]interface{}{
-		"name":                 spec.Name(),
-		"project_id":           spec.ProjectID,
-		"api_id":               spec.ApiID,
-		"version_id":           spec.VersionID,
-		"spec_id":              spec.SpecID,
-		"filename":             spec.FileName,
-		"description":          spec.Description,
-		"revision_id":          spec.RevisionID,
-		"create_time":          spec.CreateTime,
-		"revision_create_time": spec.RevisionCreateTime,
-		"revision_update_time": spec.RevisionUpdateTime,
-		"mime_type":            spec.MimeType,
-		"size_bytes":           spec.SizeInBytes,
-		"hash":                 spec.Hash,
-		"source_uri":           spec.SourceURI,
-		"labels":               labels,
-	}, nil
+func (d *Client) ListSpecs(ctx context.Context, parent names.Version, opts gorm.PageOptions) (gorm.SpecList, error) {
+	return d.Client.ListSpecs(ctx, parent, opts)
 }
 
 func (d *Client) GetSpec(ctx context.Context, name names.Spec) (*models.Spec, error) {
@@ -154,27 +35,5 @@ func (d *Client) DeleteSpec(ctx context.Context, name names.Spec) error {
 }
 
 func (d *Client) GetSpecTags(ctx context.Context, name names.Spec) ([]*models.SpecRevisionTag, error) {
-	q := d.NewQuery(gorm.SpecRevisionTagEntityName)
-	q = q.Require("ProjectID", name.ProjectID)
-	q = q.Require("ApiID", name.ApiID)
-	q = q.Require("VersionID", name.VersionID)
-	if name.SpecID != "-" {
-		q = q.Require("SpecID", name.SpecID)
-	}
-
-	var (
-		tags = make([]*models.SpecRevisionTag, 0)
-		tag  = new(models.SpecRevisionTag)
-		it   = d.Run(ctx, q)
-		err  error
-	)
-
-	for _, err = it.Next(tag); err == nil; _, err = it.Next(tag) {
-		tags = append(tags, tag)
-	}
-	if err != nil && err != iterator.Done {
-		return nil, status.Error(codes.Internal, err.Error())
-	}
-
-	return tags, nil
+	return d.Client.GetSpecTags(ctx, name)
 }
